Moved createTask actions to logistic's actions
