+++ src/redux/logistics/__tests__/taskEntityReducers.test.js
@@ -1,6 +1,6 @@
+import { createTaskSuccess } from '../../../shared/logistics/redux';
 import {
   changeDate,
-  createTaskSuccess,
   loadTasksSuccess,
 } from '../../Dispatch/actions';
 import { default as taskEntityReducers } from '../taskEntityReducers';
