+++ src/redux/logistics/__tests__/uiReducers.test.js
@@ -3,9 +3,6 @@ import {
   assignTasksRequest,
   assignTasksSuccess,
   cancelTaskSuccess,
-  createTaskFailure,
-  createTaskRequest,
-  createTaskSuccess,
   loadTaskListsFailure,
   loadTaskListsRequest,
   loadTaskListsSuccess,
@@ -17,6 +14,9 @@ import {
   loadUsersSuccess,
 } from "../../Dispatch/actions";
 import {
+  createTaskFailure,
+  createTaskRequest,
+  createTaskSuccess,
   loadToursFailure,
   loadToursSuccess,
 } from '../../../shared/logistics/redux';
